fixed migration errors
