--- tck/util/src/main/java/ee/jakarta/tck/faces/test/util/selenium/BaseArquilianRunner.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * https://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
- * which is available at https://www.apache.org/licenses/LICENSE-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the Eclipse
- * Public License v. 2.0 are satisfied: GPL-2.0 with Classpath-exception-2.0 which
- * is available at https://openjdk.java.net/legal/gplv2+ce.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 or Apache-2.0
- */
-package ee.jakarta.tck.faces.test.util.selenium;
-
-import static java.lang.Boolean.parseBoolean;
-import static org.junit.jupiter.api.extension.ConditionEvaluationResult.disabled;
-import static org.junit.jupiter.api.extension.ConditionEvaluationResult.enabled;
-
-import org.jboss.arquillian.junit5.ArquillianExtension;
-import org.junit.jupiter.api.extension.ConditionEvaluationResult;
-import org.junit.jupiter.api.extension.ExecutionCondition;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.api.extension.ExtensionContext;
-
-/**
- * Arquillian runner which triggers HtmlUnit based tests if <code>test.selenium</code> is not set to <code>true</code>.
- */
-@ExtendWith(ArquillianExtension.class)
-public class BaseArquilianRunner implements ExecutionCondition {
-
-    @Override
-    public ConditionEvaluationResult evaluateExecutionCondition(ExtensionContext context) {
-        if (parseBoolean(System.getProperty("test.selenium"))) {
-            return disabled("Test disabled because 'test.selenium' system property is not set to 'true'");
-        }
-
-        return enabled("Test enabled because 'test.selenium' system property is set to 'true'");
-    }
-}
