+++ tck/util/src/main/java/ee/jakarta/tck/faces/test/util/selenium/BaseITNG.java
@@ -15,8 +15,11 @@
  */
 package ee.jakarta.tck.faces.test.util.selenium;
 
+import static java.lang.Boolean.parseBoolean;
 import static java.lang.System.getProperty;
 import static org.jboss.shrinkwrap.api.ShrinkWrap.create;
+import static org.junit.jupiter.api.extension.ConditionEvaluationResult.disabled;
+import static org.junit.jupiter.api.extension.ConditionEvaluationResult.enabled;
 
 import java.io.File;
 import java.net.URL;
@@ -24,21 +27,37 @@ import java.time.Duration;
 import java.util.regex.Pattern;
 
 import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.shrinkwrap.api.importer.ZipImporter;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.junit.jupiter.api.extension.ConditionEvaluationResult;
+import org.junit.jupiter.api.extension.ExecutionCondition;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.ExtensionContext;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.PageFactory;
 
 /**
- * Base class for tests, which provides a certain infrastructure can be used, but does not have to be
+ * Use this for Selenium based tests.
  */
-@ExtendWith(SeleniumArquilianRunner.class)
-public class BaseITNG {
+@ExtendWith(ArquillianExtension.class)
+@TestInstance(Lifecycle.PER_CLASS)
+public abstract class BaseITNG implements ExecutionCondition {
+
+    @Override
+    public ConditionEvaluationResult evaluateExecutionCondition(ExtensionContext context) {
+        if (parseBoolean(System.getProperty("test.selenium"))) {
+            return enabled("Test enabled because 'test.selenium' system property is set to 'true'");
+        }
+
+        return disabled("Test disabled because 'test.selenium' system property is not set to 'true'");
+    }
 
     @ArquillianResource
     protected URL webUrl;
@@ -54,18 +73,18 @@ public class BaseITNG {
                 .as(WebArchive.class);
     }
 
-  @BeforeEach
-  void setUp() {
+    @BeforeEach
+    void setUp() {
         webDriver = driverPool.getOrNewInstance();
     }
 
-  @AfterEach
-  void tearDown() {
+    @AfterEach
+    void tearDown() {
         driverPool.returnInstance(webDriver);
     }
 
-  @AfterAll
-  static void afterAll() {
+    @AfterAll
+    static void afterAll() {
         driverPool.quitAll();
     }
 
@@ -79,7 +98,8 @@ public class BaseITNG {
     }
 
     /**
-     * Selenium does not automatically update the page handles if a link is clicked without ajax
+     * Selenium does not automatically update the page handles if a link is clicked
+     * without ajax
      */
     protected void updatePage() {
         ExtendedWebDriver webDriver = getWebDriver();
@@ -106,7 +126,7 @@ public class BaseITNG {
         String[] uriAndQueryString = uri.split(Pattern.quote("?"), 2);
 
         if (uriAndQueryString.length == 2) {
-            uriWithoutJsessionId += "?" + uriAndQueryString[1]; 
+            uriWithoutJsessionId += "?" + uriAndQueryString[1];
         }
 
         return uriWithoutJsessionId;
